services/video-conference/requirements.txt
fastapi>=0.104.1,<0.115
uvicorn[standard]>=0.24.0
pydantic>=2.5.0,<3
boto3>=1.34.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
